Adds perturbation filtering
Implements filtering of perturbations based on the availability of ESM2 features.

This ensures that the training process only considers perturbations for which complete feature data is available, improving the robustness and reliability of the model. A new configuration option `esm_perts_only` is added to control this behavior.

Also renames `basal_dropout_prob` to `input_dropout` for clarity.